requirements_installed.txt
adbutils==2.9.4
atomicwrites==1.4.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
ConfigArgParse==1.7.1
decorator==5.2.1
deprecation==2.1.0
emoji==2.14.1
idna==3.10
langdetect==1.0.9
lxml==6.0.1
packaging==25.0
pillow==11.3.0
PyYAML==6.0.2
requests==2.32.5
retry2==0.9.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
six==1.17.0
spintax==1.0.4
uiautomator2==3.4.1
urllib3==2.5.0
